refactor(scripts): share TypeScript AST helpers
